CAMEL-24560: Use Long for GenAiUsage token counts
Changes GenAiUsage inputTokens/outputTokens fields from Integer to Long
to avoid lossy casts and Math.toIntExact failures for large token counts
returned by OpenAI and other providers. A convenience Integer factory
overload is kept for LangChain4j/Spring AI call sites. OpenAI embeddings
producer no longer narrows usage headers to int. Micrometer counters use
increment(double) for large values. Tests and upgrade guide included.

Closes #26105

+++ components/camel-ai/camel-ai-observability/src/main/java/org/apache/camel/component/ai/observability/GenAiMicrometerSupport.java
@@ -71,7 +71,7 @@ final class GenAiMicrometerSupport implements GenAiMetricsBackend {
     }
 
     private void recordTokenCounter(
-            Integer tokens, String tokenType, GenAiObservationContext context, Throwable error) {
+            Long tokens, String tokenType, GenAiObservationContext context, Throwable error) {
         if (tokens == null || tokens <= 0) {
             return;
         }
@@ -79,7 +79,7 @@ final class GenAiMicrometerSupport implements GenAiMetricsBackend {
                 .tags(baseTags(context, error))
                 .tag(GenAiMetrics.TAG_TOKEN_TYPE, tokenType)
                 .register(meterRegistry)
-                .increment(tokens);
+                .increment(tokens.doubleValue());
     }
 
     private static Iterable<Tag> baseTags(GenAiObservationContext context, Throwable error) {
